# back/config/swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Groupomania API",
    "description": "Groupomania social media API documentation",
    "contact": {
      "email": "nssmhttb@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
...
  "paths": {
...
    "/user/refresh-token": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "This can only be done by the logged in user. Add xsrf token in **Authorize**",
        "summary": "Refresh json web token current logged user",
        "operationId": "refreshToken",
        "security": [
          {
            "cookieAuth": [],
            "refreshToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "successfully operation",
            "headers": {
              "Set-Cookie": {
                "description": "Acess json web token in cookie `access_token` (httpOnly)",
                "schema": {
                  "type": "string"
                }
              },
              "~Set-Cookie": {
                "description": "Refresh token in cookie `refresh_token` (httpOnly)",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Fetch user by id",
        "description": "This can only be done by the logged in user. Add xsrf token in **Authorize**",
        "operationId": "getUserById",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "name": "id",
            "in": "path",
            "description": "The id that needs to be fetched.",
            "required": true
          }
        ],
        "security": [
          {
            "accessToken": [],
            "xsrfToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "successfully operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserByIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/update": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update currently logged user",
        "description": "This can only be done by the logged in user. Add xsrf token in **Authorize**",
        "operationId": "updateUser",
        "security": [
          {
            "accessToken": [],
            "xsrfToken": []
          }
        ],
        "requestBody": {
          "description": "Updated user object",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successfully operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/user/delete": {
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete currently logged user",
        "description": "This can only be done by the logged in user. Add xsrf token in **Authorize**",
        "operationId": "deleteUser",
        "security": [
          {
            "accessToken": [],
            "xsrfToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "successfully opetation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpMessage"
                }
              }
            }
          },
          "400": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "expiresAt": {
            "type": "integer"
          },
          "xsrfToken": {
            "type": "string"
          }
        }
      },
      "GetUserByIdResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          }
        ],
        "properties": {
          "Posts": {
            "type": "array",
            "items": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Post"
                },
                {
                  "properties": {
                    "comment_count": {
                      "type": "integer"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "HttpErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string"
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "media": {
            "type": "string",
            "format": "binary"
          },
          "mediaType":{
            "type": "string"
          },
          "like":{
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date"
          },
          "updated_at": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "HttpMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "accessToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      },
      "refreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refrsh_token"
      },
      "xsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-xsrf-token"
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}